+++ src/webiblioteca/WebibliotecaApp.java
@@ -10,7 +10,7 @@ public class WebibliotecaApp {
     static Configs configs = new Configs();
     
     public static void main(String[] args) {
-        String version = "1.0.0 (Build 11)";
+        String version = "0.7.0 Closed Beta (Build 10)";
         
         //config check
         String env = System.getenv("APPDATA");